+++ src/main/java/com/openrca/ecommerce/model/Ship.java
@@ -19,7 +19,7 @@ public class Ship {
 
     @Id
     @Column(name = "")
-    private Integer shipId;
+    private Long shipId;
 
     @Column(name = "")
     private Integer orderId;
